mocknet improvements (#8717)
- refactored the spoon test into a few smaller methods in a class
- added some validation like 'is neard running' and 'is helper running'
- improved error handling and reporting
- improved logging
- implemented proper argument parsing in the helper
- passing test timeout and contract deploye time from spoon to helper in args
- report remainig test time in the load phase
- removed failing tx_measurement checks
- applied waclaw-python-styling


note: Currently the test passes (at least with betanet) but typically a few helper scripts die quite early in the execution. Do you, Dear Reviewers, happen to know if this is an existing issue? Either way I'll try to fix that but it would be easier if I know if this is something new or old.