fix/refactor(network::raw): rename raw::ReceivedMessage and accept message types as an arg to send_*() (#8751)
This PR changes things so that instead of calling `send_ping()` and `send_state_part_request()` to send different types of messages, we have `send_message()` and `send_routed_message()` that send direct or routed messages to the peer. This will make it easier to rewrite/add to the code in `tools/mock-node` to use a `raw::Connection` instead of mocking at the actix actor level, because we'll be able to do something like:

```
let queue: Vec<raw::Message> = ...;

match queue[0] {
  raw::Message::Direct(msg) => conn.send_message(msg),
  raw::Message::Routed(msg) => conn.send_routed_message(msg, target, ttl),
};
```

Instead of more cumbersome code that would be required at the moment, especially after adding more message types like `Block`, `BlockRequest`, etc.

This PR also adds a `RouteBackCache` to `raw::Connection`, so that we can tell if routed messages we receive are meant for us. Otherwise, as the code exists now, the `raw::Connection` will pass you any routed messages it receives as if they're meant for you, but they might not be.