fix(logging): clean up sync related log spam (#8800)
ClientActor::syncing_info() prints log messages when the bool it returns indicating whether syncing is needed doesn't match the current syncing state. The problem is that this doesn't always match up to when the client actually changes its sync status. This function is called in sync_wait_period(), which is just meant to tell how long to wait until trying to call run_sync_step() again. The result is that when all peers have disconnected from a node, it prints terrible log spam like:

```
2023-03-25T01:20:27.598443Z  WARN client: Sync: no peers available, disabling sync
2023-03-25T01:20:27.678527Z  WARN client: Sync: no peers available, disabling sync
2023-03-25T01:20:27.701984Z  WARN client: Sync: no peers available, disabling sync
...
...
```

and you get similar log spam right after node finds a peer after it had none previously:

```
2023-03-25T01:30:50.957743Z  INFO client: Sync: height: 14, peer id/height: ed25519:GyUjwR2XmFHrhsQsrW69NLKnZym6mVYqB4BLcHudv6gx/100, enabling sync
2023-03-25T01:30:50.976113Z  INFO client: Sync: height: 14, peer id/height: ed25519:GyUjwR2XmFHrhsQsrW69NLKnZym6mVYqB4BLcHudv6gx/100, enabling sync
2023-03-25T01:30:51.057412Z  INFO client: Sync: height: 14, peer id/height: ed25519:GyUjwR2XmFHrhsQsrW69NLKnZym6mVYqB4BLcHudv6gx/100, enabling sync
...
...
```

To fix it we can just have ClientActor::syncing_info() return a type that tells if we need to sync, and only print a log message in run_sync_step() when we actually change the sync status