feat: prober checks contract function call on split store node (#8737)
Make a contarct call from the prober on the legacy and the split archival nodes and check that the results are the same. 

They aren't right now and this seems to be a real issue in the split store implementation - I'll be debugging that. If anyone happens to have any idea on where this error is coming from let me know. 

```
[2023-03-15 13:21:08] INFO: Running Prober Split
[2023-03-15 13:21:08] INFO: gcloud compute instances list
[2023-03-15 13:21:09] INFO: Selected random legacy node - 35.194.243.155
[2023-03-15 13:21:10] INFO: The genesis height is 42376888. The head height is 120422174
[2023-03-15 13:21:10] INFO: Checking blocks at height 48488222
[2023-03-15 13:21:11] INFO: Checking chunks
[2023-03-15 13:21:12] ERROR: Check failed, the legacy response and the split response are different
legacy response
{'jsonrpc': '2.0', 'result': {'block_hash': 'Dok8ha8HceGRiv3MEPoxnJL2tJVxpDAo7cAc44WKSw6i', 'block_height': 120422176, 'logs': [], 'result': [48]}, 'id': 'dontcare'}
split response
{'jsonrpc': '2.0', 'error': {'name': 'INTERNAL_ERROR', 'cause': {'info': {'error_message': 'StorageError(StorageInconsistentState("cache write error"))'}, 'name': 'INTERNAL_ERROR'}, 'code': -32000, 'message': 'Server error', 'data': 'The node reached its limits. Try again later. More details: StorageError(StorageInconsistentState("cache write error"))'}, 'id': 'dontcare'}
```

The innermost erorr is `StorageInconsistentState("cache write error")`

Note to self - don't straight up compare the whole results as the block_height may be different. Only compare the 'result' field.