feat: delegate action gas params estimation (#8884)
This patch includes the final code parts of the meta-tx gas cost
estimations. It was delayed because the results were too flaky for
continuous estimations.

Specifically, this adds the estimations
1. `ActionDelegate` to estimate send + send + exec all together
2. `ActionDelegateSendNotSir`, `ActionDelegateSendSir`,
`ActionDelegateExec` to estimate each step separately 

Making these estimations stable is tricky due to the limit that only
one delegate action can be in a receipt. The check is only done in the
`send` step, so I was able to cheat my way around it for
`ActionDelegateExec`.
But I didn't find a way for the send cost estimations. 
Hence in that case we measure the cost of verifying a single action.

This results in high variation within the sample measurements, which
makes the average a poor statistic to look at. So instead, I report the
75th percentile of at least 12 measurements. The 75th percentile
was chosen over the media (50th percentile) to ensure we pick at least
a somewhat pessimistic value of the distribution.

Repeated estimations will still result in flaky results, so it's far from a 
perfect solution. But at least it removes outliers on both extremes.

In the current state, the `icount` estimations seem to be very stable,
which is what we need for continuous estimation.

For `time` based estimations, the `ActionDelegate` sometimes gets a
`HIGH-VARIANCE` or `SUBTRACTION-UNDERFLOW` uncertainty
marker. This is a common problem shared with other actions and I
don't intend to solve it in this PR. But the other other three estimations
should be stable also in `time` measurements.