feat(mock-node): mock the TCP stream instead of actix messages (#8814)
Before this change, the mock node keeps handles to the ClientActor
and ViewClientActor, and mocks the role of the PeerManagerActor. Here
we change things so that it instead just listens on a localhost socket
and sends messages to a node that has been started normally with
nearcore::start_with_config(). This has the advantage of allowing us
to in theory test/benchmark the whole node's code via only its public
API.