feat(mocknet): add a neard runner script to be run on mocknet nodes (#8983)
This adds a script that will be in charge of running neard on the machines in the mocknet test. It will download the binaries we want to run, and will change the running binary when a particular epoch height is reached. It'll also start a JSON RPC server that for now will just allow starting/stopping neard. This allows the test script to exit after the setup is finished and the test is running instead of having to keep the computer running the test online during the whole test, which can cause problems if it's interrupted for any reason.

For now the neard runner script will just be in charge of running/upgrading neard after the test has started, but in the future it would be good to move almost all initialization logic to it, so that the burden on the computer running the script is as small as possible, and interaction with the individual test nodes can happen via JSON RPC