near-vm: restore near-vm tests (#8972)
This makes sure we run all of the tests we have for the runtime. I don't particularly enjoy or appreciate the compexity of the test suite setup (e.g. we no longer need the proc macro since we only have and will ever have a single engine and compiler...), but I felt like it'd be better to restore the original functionality first and then simplify later.