jsonrpc: encode state proofs using base64 and remove deprecated field (#9091)
Change ViewStateResult::proof serialisation to base64.  This has been
the plan all along but back when proof was added¹ I forgot to follow
up with encoding change.  At the moment the proof is sent as an array
of integers which is inefficient and inconsistent with how binary data
is other JSON messages is encoded.

Furthermore, follow up with the deprecation notices: don’t include
proof field in StateItem and don’t include proof field in the response
if it hasn’t been requested.

¹ See commit faa1811d1625: ‘add proof to ViewState response’.