refactor: Separate locust workloads into different files (#9195)
This removes the need for tags and makes the start command simpler. We still can combine the workloads by passing multiple locustfiles during launch as in:
```sh
locust -H 127.0.0.1:3030 \
     -f locustfiles/ft.py,locustfiles/social.py \
     --fungible-token-wasm=$FT_CONTRACT \
     --social-db-wasm=$SOCIAL_DB_CONTRACT \
     --funding-key=$KEY \
     --headless
```

Because we don't import all workloads in a single file, there is no need to conditionally check for tags in the library modules, so I got rid of `is_tag_active` altogether.