refactor(loadtest): function call tx abstraction (#9258)
We have many transactions that are function calls under the hood.
It's about time we introduce an abstraction for that. (`class FunctionCall(Transaction)`)

Also, use a common function `create_contract_account` to create contract accounts, to standardize the message displayed in locust.

Finally, refactor the code in `ft.py` a bit to allow specialized ft contracts (SWEAT) in future PRs.