Enable TIER1 outbound by default (#9349)
This PR sets the default value for the `tier1.enable_outbound` parameter in the network config to `true`.

We are comfortable changing the default now that TIER1 has been voluntarily enabled by many validators in mainnet and used without incident for more than 2 months. At the time of this PR, more than 40 mainnet validators have TIER1 outbound enabled. 

The `enable_outbound` parameter will continue to exist, allowing for validators to opt out of the TIER1 network if desired.