fix(state-sync): Enable State Sync from GCS by default (#9398)
* fix(state-sync): Enable State Sync from GCS by default

* fix(state-sync): Enable State Sync from GCS by default

* RoutingTableView: gracefully handle missing distance vectors (#9394)

The RoutingTableView in the debug UI includes a table over the direct peers of the node:

<img width="516" alt="image" src="https://github.com/near/nearcore/assets/3241341/c8c1f0b8-1e0a-458e-a2ad-f09f20fcb57c">

Each active connection of the node has one row in the table. The last column is populated using the most recent distance vector shared by the direct peer.

In some cases there may be a direct connection for which we have not received a distance vector yet. This PR allows the debug UI to gracefully handle such cases and display "null" instead of producing an error.

* fix(state-sync): Enable State Sync from GCS by default

* fix(state-sync): Enable State Sync from GCS by default

* fix(state-sync): Enable State Sync from GCS by default

---------

Co-authored-by: Saketh Are <saketh.are@gmail.com>