feat: allow updating config.json in mirror script (#9429)
I find it useful to modify config.json-s for mirrored traffic tests. This doesn't require separate binaries which are built in >10 minutes and allows to skip setup restarts and manual ssh-ing to each node.

Here I allow to modify only state cache size because this is required for testing #9375 - we want to reduce size of 3GB cache at least. But it is easily extendable for other config parameters, I just didn't find generic way to do it.